# courses/backend/node/week2/api/src/openapi.yml
openapi: 3.0.0

info:
  title: Snippets API
  version: 1.0.0
  description: API for managing snippets and tags

servers:
  - url: http://localhost:3000/api


tags:
  - name: Snippets
    description: All snippet related endpoints

  - name: Tags
    description: All tag related endpoints


paths:

  # =========================
  # SNIPPETS
  # =========================

  /snippets:
    get:
      tags: [Snippets]
      summary: Get all snippets
      parameters:
        - in: query
          name: sort
          schema:
            type: string
            example: "created_at desc"
      responses:
        "200":
          description: Success
        "400":
          description: Bad request
        "500":
          description: Server error

    post:
      tags: [Snippets]
      summary: Create a snippet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SnippetInput"
            example:
              title: "My snippet"
              contents: "console.log('hello')"
              user_id: 1
      responses:
        "201":
          description: Created
        "400":
          description: Bad request
        "500":
          description: Server error


  /snippets/search:
    get:
      tags: [Snippets]
      summary: Search snippets
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "400":
          description: Bad request
        "500":
          description: Server error


  /snippets/public:
    get:
      tags: [Snippets]
      summary: Get public snippets
      responses:
        "200":
          description: Success
        "500":
          description: Server error


  /snippets/{id}:
    get:
      tags: [Snippets]
      summary: Get snippet by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
        "404":
          description: Not found

    put:
      tags: [Snippets]
      summary: Update snippet
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SnippetInput"
            example:
              title: "Updated title"
              contents: "updated code"
              user_id: 1
      responses:
        "200":
          description: Updated
        "404":
          description: Not found

    delete:
      tags: [Snippets]
      summary: Delete snippet
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
        "404":
          description: Not found


  # =========================
  # TAGS
  # =========================

  /tags:
    get:
      tags: [Tags]
      summary: Get all tags
      responses:
        "200":
          description: Success

    post:
      tags: [Tags]
      summary: Create tag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagInput"
            example:
              name: "javascript"
      responses:
        "201":
          description: Created
        "400":
          description: Missing name


  /tags/{id}:
    get:
      tags: [Tags]
      summary: Get tag by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
        "404":
          description: Not found

    put:
      tags: [Tags]
      summary: Update tag
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagInput"
            example:
              name: "nodejs"

      responses:
        "200":
          description: Updated
        "400":
          description: Bad request
        "404":
          description: Not found

    delete:
      tags: [Tags]
      summary: Delete tag
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
        "404":
          description: Not found


components:

  schemas:

    Snippet:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        contents:
          type: string
        user_id:
          type: integer
        is_private:
          type: boolean
        created_at:
          type: string

    SnippetInput:
      type: object
      required:
        - title
        - contents
        - user_id
      properties:
        title:
          type: string
        contents:
          type: string
        user_id:
          type: integer


    Tag:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string

    TagInput:
      type: object
      required:
        - name
      properties:
        name:
          type: string